requirements.txt
decord
einops
compress_pickle
pytorch_msssim
opencv-python
scikit-video
tqdm
ninja
torchac
timm
python-dotenv
pandas
matplotlib
